requirements.txt
pandas==2.2.2
waitress==3.0.0
Flask==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.1